tests: make toolchain tests run in the same build instead of bazel-in-bazel integration tests (#2095)
This changes the //tests/toolchains tests to run in the same Bazel build
instance instead of
being bazel-in-bazel integration tests. This is done by using a
transition to set
the python version to match the desired toolchain.

This makes running the tests **much** cheaper -- there were 37 sub-bazel
processes being
started (one for each python version), which was very expensive.
Debugging is also much easier because they aren't part of a
bazel-in-bazel invocation.