test(toolchain): use flag_values in test platform defs (#2106)
This PR uses the `flag_values` from the platform definitions
for the toolchains so that in the future we can distinguish
between `musl` and `glibc` toolchains in our tests.

For now the change is no-op.

As part of this change we are also registering the coverage
tools so that we can run `bazel coverage` with no errors.

See comment on #2095.