feat(rules): add PyExecutableInfo
The PyExecutableInfo provider exposes executable-specific information
that isn't easily accessible outside the target. The main purpose of
this provider is to facilitate packaging a binary or deriving a new
binary based upon the original.

Within rules_python, this will be used to pull the zip-building logic
out of executables and into separate rules.