feat(rules): add PyExecutableInfo (#2166)
The PyExecutableInfo provider exposes executable-specific information
that isn't easily accessible outside the target. The main purpose of
this provider is to facilitate packaging a binary or deriving a new
binary based upon the original.

Within rules_python, this will be used to pull the zip-building logic
out of executables and into separate rules. Within Google, this will be
used for a similar "package a binary" tool.

Along the way:
* Add runfiles references to sphinx inventory