feat(rules): add build_data_file field to PyExecutableInfo
PyExecutableInfo was added in https://github.com/bazelbuild/rules_python/pull/2166 with the field `runfiles_without_exe` that intentionally excludes unrelated files. But, within Google we found that we need to pick up the file with build stamp information at that stage as well, so it is being additionally exposed.