fix: allow detecting if `--precompile_source_retention` was specified on the command line (#2192)
This defaults the `--precompile_source_retention` flag to a new value,
`auto`. The effective default remains the same (`keep_source`).

Having a separate value as the default allows detecting if the value was
specified on the command line, which allows distinguishing between "pick
for me" vs "I specifically want this behavior".