feat(rules): add build_data_file field to PyExecutableInfo (#2181)
PyExecutableInfo was added in
https://github.com/bazelbuild/rules_python/pull/2166 with the field
`runfiles_without_exe` that intentionally excludes files that are
specific to
that target/executable, such as the build data file (which may contain
the target name,
or other target-specific information).

However, consuming tools (such as ones used within Google) may need to
derive a file from
that build data, override it completely, or be happy with its content as
is. To aid that
case, expose it via PyExecutableInfo.