fix: allow disabling exec_tools toolchain from looking up an interpreter (#2194)
This allow the exec_interpreter attribute to propagate None even though
it's a label
with a default value. Such attributes can't _directly_ be set to None
because None
means to use the default. To work around that, a sentinel target is
used, which allows
breaking the dependency on the default and triggering the rule to use
None instead.

A null exec interpreter is necessary in environments that don't provide
a separate Python
interpreter.