fix(sphinx): Support python 3.9 in Sphinx rules (#2208)
The implementation of the Sphinx related rules uses language features
and types only available in python > 3.9. To the best of my
understanding, `rules_python` aims to support python3.9 and therefore
this patch uses `typing-extensions` instead of the built-in types.

After this change, one can use the `sphinx_build_binary` to create
Sphinx docs using 3.9 runtime.

This fixes #2207