tests: make precompile tests pass when other toolchains are defined (#2213)
This makes the precompile_tests pass when the environment defines custom
toolchains that
don't match what rules_python defines in its dev environment. This keeps
the tests
independent of whatever the user's environment is.

Basically, the tests rely on a fake Python version 4.5 toolchain for the
precompiler being
defined. They pass today because they fallback to a real toolchain (as
setup by
MODULE.bazel), which doesn't have any version constraints on it. In
comparison, within
Google, there is no "default" toolchain, so the tests fail to find what
they need.

To fix, explicitly define a fake precompiler toolchain and tell the test
to use it.

Along the way:
* Also force `--allow_unresolved_symlinks=true` in the tests. This flag
isn't enabled in
  certain environments, but is implicitly relied upon by the
`current_interpreter_executable` rule when a platform runtime is used
(as they are in
  the tests).
* Move the Python testing toolchains to support/py_toolchains, to match
where the cc
  testing toolchains were moved.