refactor: move hermetic Python runtime setup into macro (#2221)
This stems from conversation in #2216. Moving the logic into a
loading-phase macro makes it
easier to understand how a repo with a python-build-standalone extracted
archive has
its toolchain defined. This also makes it easier to compare with the
local runtime repo
setup to try and find commonality in the future.

Along the way:
* Remove some Bazel 5 compatibility code: coverage_tool is now always
passed to py_runtime.
(Bazel 5 support was dropped earlier this year and we stopped testing
with it; it's
  likely already broken with Bazel 5).
* Add `semver.to_dict`, which makes it easier to pass into format()
functions.