fix: Prefix bootstrap file to prevent pytest reentrance (#2230)
Pytest will, by default, discover tests in files named `test_*.py`. The
existing behaviour of the bootstrap setup is to create a file named e.g.
`test_foo_stage2_bootstrap.py` for a target named `test_foo`. This
causes pytest to reenter, leading to a crash.

To fix, prepend the generated stage2 file with an underscore so that
pytest doesn't match it.

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>