chore: delete stale bot (#2264)
The repo is actively maintained, but we've had a bunch of relevant
issues be auto-closed by the bot. I haven't seen other active rulesets
adopt this, so I don't think it makes sense to be the outlier.

Note, I introduced this configuration.

https://github.com/bazelbuild/rules_python/issues/385 as an example
where users have to make comments to prevent things being marked stale.