Conditionally remove builtin PyInfo and other providers
Use bazel_features to detect if PyInfo is still present in Bazel and
stop using it, if it's not.

Bazel throws an error while compiling bzl files if there is a
reference to a top-level symbol that doesn't exist anymore.
bazel_features provide a mechanism that detects presence of top-level
symbols without throwing an exception. It either returns the symbol or None.

bazel_features support not only Bazel 8, but also older versions of
Bazel that have --incompatible_autoload_externally cherry-picked. If the
flag is enabled with "@rules_python" or "-@rules_python" the provider is
removed from Bazel.

Remove also PyRuntimeInfo and PyCcLinkParamsProvider.