feat(toolchains): default py_runtime.implementation_name to cpython (#2272)
This defaults `py_runtime.implementation_name` to `cpython`, the value
that is most likely
to be used. This is done not only because it's the most likely value,
but because of how
it affects whether precompiling works.

Within py_runtime, automatically computing `pyc_tag` requires both
`implementation_name`
and Python version information. With upcoming changes to make
`--python_version` always
set to the correct value, it'll make enabling precompiling easier,
especially if there
are custom toolchains defined.