fix(sphinxdocs): allow using sphinx_stardoc with bzlmod
Using the sphinx_stardoc rule doesn't work with bzlmod because it
refers to the stardoc repo, which was a dev dependency. This means,
even if a user's root module has visibility to stardoc, rules_python
doesn't.

To fix, make stardoc a non-dev dependency. With bzlmod semantics,
stardoc won't actually be loaded unless a user depends on it by
using sphinxdocs.