fix(sphinxdocs),deps: allow using sphinx_stardoc with bzlmod; add stardoc 0.6.2 as dependency (#2295)
Using the sphinx_stardoc rule doesn't work with bzlmod because it refers
to the stardoc repo, which was a dev dependency. This means, even if a
user's root module has visibility to stardoc, rules_python doesn't.

To fix, make stardoc a non-dev dependency. With bzlmod semantics,
stardoc won't actually be loaded unless a user depends on it by using
sphinxdocs.

I ran into this while trying to port rules_testing over to sphinxdocs.
The pigweed project
recently started using sphinxdocs and didn't run into this problem, but
I'm pretty
sure that's because they're using workspace still, which doesn't have
the visibility
constraints that bzlmod does.