fix(precompiling): only add pyc to default outputs if precompiling explicitly enabled for target (#2307)
This fixes a bug where precompiled files were *always* being added to
the default outputs
of a target. The intent is they are only added to a target's default
outputs if the
target explicitly opted into precompiling. This went unnoticed because
the exec tools
toolchain is still disabled by default, so the test that verifies a
basic py_binary's
default outputs didn't generate implicit pyc files. It was introduced
when fixing the
pyc collection bug.

To fix, only add the pyc files deemed "required", not all the pyc files
generated.

Also added a test to capture this case to the precompile tests.