test(bzlmod): refactor pip_parse to make it more testable (#2314)
This PR introduces a new `parse_modules` function in the
`pypi/extension.bzl`
code to mimic the structure of the `python` extension and to make it
easier to
write unit tests against the extension itself. I have also written a few
unit
tests to verify the generic structure.

Work towards #2268.