fix(rules): remove rules_python --incompatible_python_disallow_native_rules checking (#2327)
When --incompatible_python_disallow_native_rules is enabled, all the
core rules fail with
an error that rules_python should be used. This is incorrect, since the
rules_python rules
are being used. What's happening is
https://github.com/bazelbuild/rules_python/pull/2257
removed the magic migration tag when pystar is enabled, but the code to
check the tag
was present wasn't removed. This went unnoticed because our CI doesn't
set the migration
flag.

To fix, remove the validation logic entirely. If we're in the
rules_python implementation,
then there is not need to perform this validation. It was just something
copy/pasted from
the original code from Bazel itself.

Also update the bazelrc to always set
--incompatible_python_disallow_native_rules.

Fixes https://github.com/bazelbuild/rules_python/issues/2326
Fixes https://github.com/bazelbuild/rules_python/issues/1645