chore: bump twine dependencies and cleanup for 1.0 (#2333)
This change updates the `requirements.txt` for all of the platforms for
the `twine` component that is used in wheel publishing rules. Before
that
the deps would be maintained by `dependabot.txt` and it seems that it
would
constantly exclude the `requirements.txt` from the updated files, this
means
that the `requirements.txt` (used on Linux) has gone out of sync and it
would
be better to use a different mechanism to keep them in sync.

Hence the proposal is to:
- Use `uv` to sync the requirements.
- Add `requirements_linux.txt` to explicitly highlight that the file is
for
  Linux.
- Keep `requirements.txt` but note that it is deprecated.
- Add `requirements_universal.txt` in case people want to use a single
  requirements file.
- Bump the dependencies to the latest versions and bring them back in
sync.

Whilst at it I had to:
- support extra parameters in the locking format and allow setting
visibility
- chore: add `rules_multirun` as a `dev_dependency`
- refactor: `uv pip compile` requirements for tools/publish
- chore: bump uv to 0.4.25