fix(bzlmod): ensure that config_settings are generated for all available versions
I think that #2253 has introduced an issue that went unnoticed because
we did not have an integration test checking this. The main issue is
that config settings of `is_python_3.x.y` were being generated only for
registered toolchains and that is not necessarily the desired behaviour
-- we may want to use the config settings for more values, so coupling
that was a mistake.