refactor: fold bazel-specific code into main files (#2358)
This removes most of the `*_bazel.bzl` files and either renames them or
folds them
into the "regular" files. (common_bazel.bzl and py_executable_bazel.bzl
turned out
to be a bit more involved, so are left for a future change).

This is part of simplifying the implementation and removing the
complicated/convoluted
"plugin" style design that was inherited from the original Bazel code to
accommodate
various Google changes.