feat: add attrs/fields for the runtime's ABI flags (#2390)
This adds attributes and fields for storing the runtimes ABI flags
value, i.e. `sys.abiflags`.

For freethreaded interpreters, the abi flags contain `t`, which is used
creating e.g.
virtualenvs.

The attribute can be directly specified, or if it's not, will be
computed based on the
`--py_freethreaded` flag.