fix: skip precompiling if using Bazel-builtin PyRuntimeInfo (#2394)
When the `@bazel_tools//python/tools:python_autodetecting` toolchain is
used, it returns
the Bazel-builtin PyRuntimeInfo provider. Because this provider doesn't
support the
additional fields needed for precompiling (`pyc_tag`, among others), it
would result in
an attribute error.

To fix, detect the absence of the `pyc_tag` attribute and return early,
as is done when
the pyc_tag field is empty.

Fixes https://github.com/bazelbuild/rules_python/issues/2364